WEB-157: [refactor] extract helper functions
